requirements.txt
# 多模态情绪识别系统依赖
# 安装命令: pip install -r requirements.txt

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# 文本处理
transformers>=4.20.0
huggingface-hub

# Web框架
flask>=2.0.0
flask-cors

# 图像和视频处理
opencv-python>=4.5.0
moviepy>=1.0.3
Pillow>=8.0.0
imageio-ffmpeg

# 语音识别
openai-whisper

# 数据处理
numpy>=1.20.0
scipy>=1.6.0
pandas>=1.2.0
scikit-learn>=0.24.0

# 可视化
matplotlib>=3.3.0
seaborn>=0.11.0

# 工具
tqdm>=4.60.0

# 可选：如果需要运行Web前端
# chart.js (通过CDN引入，无需pip安装)
